# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion des articles",
        "description": "API pour gérer les articles et leur stock",
        "version": "1.0"
    },
    "paths": {
...
        "/apiv1/clients": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Récupère la liste de tous les clients",
                "operationId": "5a2023c1d79734979bf087a127922922",
                "responses": {
                    "200": {
                        "description": "Liste des clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des clients"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun client trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pas de clients"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clients": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Récupère la liste des clients avec ou sans comptes",
                "operationId": "9fbaa4228c73b1ac004a423578684f7d",
                "parameters": [
                    {
                        "name": "comptes",
                        "in": "query",
                        "description": "Filtrer les clients par comptes. Utiliser 'oui' pour les clients avec comptes, 'non' pour sans comptes",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "oui",
                                "non"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des clients"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Paramètre invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paramètre invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clients/ActiveStatus": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Récupère la liste des clients en fonction du statut actif de leur compte",
                "operationId": "ffbd10691187b35689d1379af1eabfff",
                "parameters": [
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filtrer les clients par statut actif du compte. Utiliser 'oui' pour les clients actifs, 'non' pour inactifs",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "oui",
                                "non"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Client"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des clients"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Paramètre invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paramètre invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clients/telephone": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Recherche un client par numéro de téléphone",
                "operationId": "825d7e44cc3b3114e7d9768f6996d7f4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "telephone": {
                                        "type": "string",
                                        "example": "0612345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client trouvé ou non",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Client"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Client trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "411": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 411
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Client": {
                "title": "Client",
                "description": "Client schema",
                "properties": {
                    "id": {
                        "description": "ID du client",
                        "type": "integer",
                        "example": 1
                    },
                    "surname": {
                        "description": "Nom de famille du client",
                        "type": "string",
                        "example": "Dupont"
                    },
                    "adresse": {
                        "description": "Adresse du client",
                        "type": "string",
                        "example": "123 Rue Example"
                    },
                    "telephone": {
                        "description": "Numéro de téléphone du client",
                        "type": "string",
                        "example": "0612345678"
                    },
                    "user": {
                        "description": "Informations de l'utilisateur associé",
                        "properties": {
                            "login": {
                                "type": "string",
                                "example": "client_login"
                            },
                            "password": {
                                "type": "string",
                                "example": "Password123!"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT authorization header using the Bearer scheme."
            }
        }
    },
...
    "security": [
        {
            "BearerToken": []
        }
    ]
}